Removing dist/ folder from the build
Addresses #2

Git is designed for source files, not built files. There are many
reasons but the most obvious one is the history for a built file isn't
meaningful in a repository.

This uses the common pattern of a prepublish command which will be run
any time before `npm publish` is executed.

This bit me while making another pull request so I figured I'd get both
birds with one stone right now. This removes the manual human coupling
of needing to run a specific script before publishing.